# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "拾取、清洗、归档你关心的内容。把 RSS、网页、聊天机器人里的信息统一成同一个阅读流。",
        "title": "NewsGlean API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
        "/entry/list": {
            "get": {
                "description": "分页列出条目，可按渠道过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "列出条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "按渠道过滤",
                        "name": "source_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.EntryListResponse"
                        }
                    }
                }
            }
        },
        "/entry/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "获取单个条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.EntryDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "遍历所有启用的渠道拉取新内容并写库（三层去重）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refresh"
                ],
                "summary": "手动触发一轮采集",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.RefreshResult"
                        }
                    }
                }
            }
        },
        "/source": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "列出渠道实例",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SourceListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "校验配置后入库（走适配器 Validate，失败不入库）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "新增渠道实例",
                "parameters": [
                    {
                        "description": "渠道参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.SourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SourceDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "获取单个渠道实例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道实例ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SourceDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "更新渠道实例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道实例ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "渠道参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.SourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SourceDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "删除渠道实例（软删除）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道实例ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.RefreshResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "错误信息",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inserted": {
                    "description": "新入库条目数",
                    "type": "integer"
                },
                "skipped": {
                    "description": "去重跳过的条目数",
                    "type": "integer"
                },
                "sources": {
                    "description": "处理的渠道数",
                    "type": "integer"
                }
            }
        },
        "server.EntryDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "content": {
                    "description": "正文",
                    "type": "string"
                },
                "content_type": {
                    "description": "内容类型",
                    "type": "string"
                },
                "extra": {
                    "description": "渠道特有字段",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "fetched_at": {
                    "description": "抓取时间（RFC3339）",
                    "type": "string"
                },
                "guid": {
                    "description": "渠道自带稳定ID",
                    "type": "string"
                },
                "id": {
                    "description": "条目ID",
                    "type": "integer"
                },
                "published_at": {
                    "description": "发布时间（RFC3339）",
                    "type": "string"
                },
                "source_id": {
                    "description": "归属渠道实例ID",
                    "type": "integer"
                },
                "summary": {
                    "description": "摘要",
                    "type": "string"
                },
                "tags": {
                    "description": "标签",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "url": {
                    "description": "规范化后的链接",
                    "type": "string"
                }
            }
        },
        "server.EntryListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "条目列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.EntryDTO"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "server.SourceDTO": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "渠道配置 JSON 对象",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "fail_count": {
                    "description": "连续失败次数",
                    "type": "integer"
                },
                "id": {
                    "description": "渠道实例ID",
                    "type": "integer"
                },
                "interval": {
                    "description": "刷新间隔（秒）",
                    "type": "integer"
                },
                "last_error": {
                    "description": "最近一次错误",
                    "type": "string"
                },
                "name": {
                    "description": "显示名",
                    "type": "string"
                },
                "type": {
                    "description": "渠道类型标识",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "server.SourceListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "渠道列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.SourceDTO"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "server.SourceRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "渠道配置 JSON 对象",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "interval": {
                    "description": "刷新间隔（秒）",
                    "type": "integer"
                },
                "name": {
                    "description": "显示名",
                    "type": "string"
                },
                "type": {
                    "description": "渠道类型标识",
                    "type": "string"
                }
            }
        }
    }
}